Replaced Expression<Func<>> to Func<> in the extension methods for Length and RegularExpression Validators. Removed unnecesary casts in LengthValidator Now Regex validator preserves the Regex object in applicable cases Replaced spaces with tabs in my new code Added unit tests for the new extension methods

+++ src/FluentValidation/Validators/LengthValidator.cs
@@ -26,9 +26,8 @@ namespace FluentValidation.Validators {
 		public int Min { get; private set; }
 		public int Max { get; private set; }
 
-        public Func<object, int> MinFunc { get; set; }
-
-        public Func<object, int> MaxFunc { get; set; }
+		public Func<object, int> MinFunc { get; set; }
+		public Func<object, int> MaxFunc { get; set; }
 
 		public LengthValidator(int min, int max) : this(min, max, () => Messages.length_error) {
 		}
@@ -42,23 +41,23 @@ namespace FluentValidation.Validators {
 			}
 		}
 
-        public LengthValidator(Func<object, int> min, Func<object, int> max)
-            : this(min, max, () => Messages.length_error) {
-        }
+		public LengthValidator(Func<object, int> min, Func<object, int> max)
+			: this(min, max, () => Messages.length_error) {
+		}
 
-        public LengthValidator(Func<object, int> min, Func<object, int> max, Expression<Func<string>> errorMessageResourceSelector) : base(errorMessageResourceSelector) {
-            MaxFunc = max;
-            MinFunc = min;
-        }
+		public LengthValidator(Func<object, int> min, Func<object, int> max, Expression<Func<string>> errorMessageResourceSelector) : base(errorMessageResourceSelector) {
+			MaxFunc = max;
+			MinFunc = min;
+		}
 
 		protected override bool IsValid(PropertyValidatorContext context) {
 			if (context.PropertyValue == null) return true;
 
-            if (MaxFunc != null && MinFunc != null)
-            {
-                Max = (int)MaxFunc(context.Instance);
-                Min = (int)MinFunc(context.Instance);
-            }
+			if (MaxFunc != null && MinFunc != null)
+			{
+				Max = MaxFunc(context.Instance);
+				Min = MinFunc(context.Instance);
+			}
 
 			int length = context.PropertyValue.ToString().Length;
 
@@ -80,53 +79,53 @@ namespace FluentValidation.Validators {
 			
 		}
 
-        public ExactLengthValidator(Func<object, int> length)
-            : base(length, length, () => Messages.exact_length_error) {
+		public ExactLengthValidator(Func<object, int> length)
+			: base(length, length, () => Messages.exact_length_error) {
 
-        }
+		}
 	}
 
-    public class MaximumLengthValidator : LengthValidator {
-        public MaximumLengthValidator(int max) : this(max, () => Messages.length_error) {
+	public class MaximumLengthValidator : LengthValidator {
+		public MaximumLengthValidator(int max) : this(max, () => Messages.length_error) {
 
-        }
+		}
 
-        public MaximumLengthValidator(int max, Expression<Func<string>> errorMessageResourceSelector)
-            : base(0, max, errorMessageResourceSelector) {
+		public MaximumLengthValidator(int max, Expression<Func<string>> errorMessageResourceSelector)
+			: base(0, max, errorMessageResourceSelector) {
 
-        }
+		}
 
-        public MaximumLengthValidator(Func<object, int> max) : 
-            this(max, () => Messages.length_error) { 
+		public MaximumLengthValidator(Func<object, int> max) : 
+			this(max, () => Messages.length_error) { 
 
-        }
+		}
 
-        public MaximumLengthValidator(Func<object, int> max, Expression<Func<string>> errorMessageResourceSelector)
-            : base(obj => 0, max, errorMessageResourceSelector) {
+		public MaximumLengthValidator(Func<object, int> max, Expression<Func<string>> errorMessageResourceSelector)
+			: base(obj => 0, max, errorMessageResourceSelector) {
 
-        }
-    }
+		}
+	}
 
-    public class MinimumLengthValidator : LengthValidator {
-        public MinimumLengthValidator(int min) : this(min, () => Messages.length_error) {
+	public class MinimumLengthValidator : LengthValidator {
+		public MinimumLengthValidator(int min) : this(min, () => Messages.length_error) {
 
-        }
+		}
 
-        public MinimumLengthValidator(int min, Expression<Func<string>> errorMessageResourceSelector) 
-            : base(min, -1, errorMessageResourceSelector) {
+		public MinimumLengthValidator(int min, Expression<Func<string>> errorMessageResourceSelector) 
+			: base(min, -1, errorMessageResourceSelector) {
 
-        }
+		}
 
-        public MinimumLengthValidator(Func<object, int> min)
-            : this(min, () => Messages.length_error) {
+		public MinimumLengthValidator(Func<object, int> min)
+			: this(min, () => Messages.length_error) {
 
-        }
+		}
 
-        public MinimumLengthValidator(Func<object, int> min, Expression<Func<string>> errorMessageResourceSelector)
-            : base(min, obj => -1, errorMessageResourceSelector) {
+		public MinimumLengthValidator(Func<object, int> min, Expression<Func<string>> errorMessageResourceSelector)
+			: base(min, obj => -1, errorMessageResourceSelector) {
 
-        }
-    }
+		}
+	}
 
 	public interface ILengthValidator : IPropertyValidator {
 		int Min { get; }
